Revert "fixed comments"
This reverts commit c396588aa671dd121617197be0e600a93a8b1c90.
