+++ packages/main/src/Calendar.hbs
@@ -54,7 +54,6 @@
 			id="{{_id}}-head"
 			.primaryCalendarType="{{_primaryCalendarType}}"
 			.secondaryCalendarType="{{secondaryCalendarType}}"
-			.secondaryCalendarMonthButtonInfo="{{displayedSecondaryMonthText}}"
 			timestamp="{{_timestamp}}"
 			.isPrevButtonDisabled="{{_previousButtonDisabled}}"
 			.isNextButtonDisabled="{{_nextButtonDisabled}}"
